Now you can combine views using the source menu.  The source menu now works in two ways.  First, the old point and click with 'A' to select a single source.  And now, Second, choose the sources you want to see with the 'B' button.  They'll click from off to on and back.  when done selecting click on the back button or press 'Home'.
Note: The All Plugins source option has been removed and choosing either nand button on mutiple selection loads the nand you last used (real or emu)

The categories_v4.ini had to also be slightly modified to work with this. It will now contain a single 'GENERAL' section which contains the number of categories and the name of each.  A suggestion, I would use the first 20 for bostonbc's genres and ratings and the next ten or so for movie types(mplayerce) and the next ten or so for misc.

When using more than one source the partition selection and emu nand settings in config are disabled for now.

The source button on the main screen still works the same including the d-pad option.

Most of the credit for this goes to Fix94 for setting wiiflow up for this a long time ago.  I just took the 80% of work he'd already did and finished by getting it to work with the source menu.  Note there may be some issues.

git-svn-id: https://svn.code.sf.net/p/wiiflow-lite/code/trunk@852 81555f4c-30bf-4792-b450-80a1b71d4377